requirements.txt
pytest
playwright
pytest-html
allure-pytest
pytest-xdist
requests
